+++ corekit/src/main/java/com/midtrans/sdk/corekit/models/PaymentDetails.java
@@ -48,15 +48,15 @@ public class PaymentDetails {
         return transactionDetails != null ? transactionDetails.getAmount() : 0;
     }
 
-    public void setTotalAmount(long totalAmount) {
+    public void setTotalAmount(double totalAmount) {
         if (transactionDetails != null) {
-            transactionDetails.setAmount((int) totalAmount);
+            transactionDetails.setAmount(totalAmount);
         }
     }
 
-    public void changePaymentDetails(List<ItemDetails> newItemDetails, long newTotalAmount) {
+    public void changePaymentDetails(List<ItemDetails> newItemDetails, double newTotalAmount) {
         if (transactionDetails != null) {
-            transactionDetails.setAmount((int) newTotalAmount);
+            transactionDetails.setAmount(newTotalAmount);
             this.itemDetailsList = newItemDetails;
         }
     }
